requirements.txt
asgiref==3.3.1
coverage==5.3
Django==3.1.3
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-sql-explorer==2.0
django-webshell==0.3.0
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1